fix: recon hdobs gridded_stats
