+++ src/tropycal/tracks/plot.py
@@ -2094,228 +2094,8 @@ class TrackPlot(Plot):
                                              color='k'),
                              transform=ccrs.PlateCarree(), clip_on=True)
 
-    def plot_gridded(self, xcoord, ycoord, zcoord, varname='type', VEC_FLAG=False, domain="north_atlantic", ax=None, prop={}, map_prop={}):
-        r"""
-        Creates a plot of a single storm track.
-
-        Parameters
-        ----------
-        storm : str, tuple or dict
-            Requested storm. Can be either string of storm ID (e.g., "AL052019"), tuple with storm name and year (e.g., ("Matthew",2016)), or a dict entry.
-        domain : str
-            Domain for the plot. Default is TrackDataset basin. Can be one of the following:
-            "north_atlantic" - North Atlantic Ocean basin
-            "pacific" - East/Central Pacific Ocean basin
-            "lonW/lonE/latS/latN" - Custom plot domain
-        plot_all_dots : bool
-            Whether to plot dots for all observations along the track. If false, dots will be plotted every 6 hours. Default is false.
-        ax : axes
-            Instance of axes to plot on. If none, one will be generated. Default is none.
-        prop : dict
-            Property of storm track lines.
-        map_prop : dict
-            Property of cartopy map.
-        """
-
-        # Set default properties
-        default_prop = {'cmap': 'category', 'levels': None,
-                        'left_title': '', 'right_title': 'All storms',
-                        'plot_values': False, 'values_size': None}
-
-        # Initialize plot
-        prop = self.add_prop(prop, default_prop)
-        map_prop = self.add_prop(map_prop, default_map_prop)
-        self.plot_init(ax, map_prop)
-
-        # Determine if contour levels are automatically generated
-        auto_levels = True if prop['levels'] is None or prop['levels'] == [
-        ] else False
-
-        # Plot domain
-        bound_w, bound_e, bound_s, bound_n = self.set_projection(domain)
-
-        # Plot parallels and meridians
-        # This is currently not supported for all cartopy projections.
-        try:
-            self.plot_lat_lon_lines([bound_w, bound_e, bound_s, bound_n], check_prop=True)
-        except:
-            pass
-
-        # --------------------------------------------------------------------------------------
-
-        if VEC_FLAG:
-            vecmag = np.hypot(*zcoord)
-            if prop['levels'] is None:
-                prop['levels'] = (np.nanmin(vecmag), np.nanmax(vecmag))
-        elif prop['levels'] is None:
-            prop['levels'] = (np.nanmin(zcoord), np.nanmax(zcoord))
-        cmap, clevs = get_cmap_levels(varname, prop['cmap'], prop['levels'])
-
-        # Generate contourf levels
-        if len(clevs) == 2:
-            y0 = min(clevs)
-            y1 = max(clevs)
-            dy = (y1 - y0) / 8
-            scalemag = int(np.log(dy) / np.log(10))
-            dy_scaled = dy * 10**-scalemag
-            dc = min([1, 2, 5, 10], key=lambda x: abs(x - dy_scaled))
-            c0 = np.ceil(y0 / dc * 10**-scalemag) * dc * 10**scalemag
-            c1 = np.floor(y1 / dc * 10**-scalemag) * dc * 10**scalemag
-            clevs = np.arange(c0, c1 + dc, dc)
-
-        if varname == 'vmax' and prop['cmap'] == 'category':
-            vmin = min(clevs)
-            vmax = max(clevs)
-        else:
-            vmin = min(prop['levels'])
-            vmax = max(prop['levels'])
-
-        # For difference/change plots with automatically generated contour levels, ensure that 0 is in the middle
-        if auto_levels:
-            if varname in ['dvmax_dt', 'dmslp_dt'] or '\n' in prop['title_R']:
-                max_val = np.max([np.abs(vmin), vmax])
-                vmin = np.round(max_val * -1.0, 2)
-                vmax = np.round(max_val * 1.0, 2)
-                clevs = [vmin, np.round(vmin * 0.5, 2),
-                         0, np.round(vmax * 0.5, 2), vmax]
-
-        if len(xcoord.shape) and len(ycoord.shape) == 1:
-            xcoord, ycoord = np.meshgrid(xcoord, ycoord)
-
-        if VEC_FLAG:
-            binsize = abs(xcoord[0, 0] - xcoord[0, 1])
-            norm = mcolors.Normalize(vmin=vmin, vmax=vmax)
-            cbmap = self.ax.pcolor(xcoord, ycoord, vecmag[:-1, :-1], cmap=cmap, vmin=min(clevs), vmax=max(clevs),
-                                   transform=ccrs.PlateCarree())
-            zcoord = zcoord / vecmag * binsize
-            x_center = (xcoord[:-1, :-1] + xcoord[1:, 1:]) * .5
-            y_center = (ycoord[:-1, :-1] + ycoord[1:, 1:]) * .5
-            u = zcoord[0][:-1, :-1]
-            v = zcoord[1][:-1, :-1]
-            if not prop['plot_values']:
-                self.ax.quiver(x_center, y_center, u, v, color='w', alpha=0.6, transform=ccrs.PlateCarree(),
-                               pivot='mid', width=.001 * binsize, headwidth=3.5, headlength=4.5, headaxislength=4)
-            zcoord = vecmag
-
-        else:
-            print('--> Generating plot')
-            # if varname=='date' and prop['smooth'] is not None:
-            #    zcoord[np.isnan(zcoord)]=0
-            #    zcoord=gfilt(zcoord,sigma=prop['smooth'])
-            #    zcoord[zcoord<min(clevs)]=np.nan
-
-            if prop['cmap'] == 'category' and varname == 'vmax':
-                norm = mcolors.BoundaryNorm(clevs, cmap.N)
-                cbmap = self.ax.pcolor(xcoord, ycoord, zcoord[:-1, :-1], cmap=cmap, norm=norm,
-                                       transform=ccrs.PlateCarree())
-            else:
-                norm = mcolors.Normalize(vmin=vmin, vmax=vmax)
-                cbmap = self.ax.pcolor(xcoord, ycoord, zcoord[:-1, :-1], cmap=cmap, norm=norm,
-                                       transform=ccrs.PlateCarree())
-        if prop['plot_values']:
-            binsize = abs(xcoord[0, 0] - xcoord[0, 1])
-            x_center = (xcoord[:-1, :-1] + xcoord[1:, 1:]) * .5
-            y_center = (ycoord[:-1, :-1] + ycoord[1:, 1:]) * .5
-            xs = x_center.flatten(order='C')
-            ys = y_center.flatten(order='C')
-            zs = zcoord[:-1, :-1].flatten(order='C')
-            if prop['values_size'] is None:
-                fs = binsize * 4
-            else:
-                fs = prop['values_size']
-            for xtext, ytext, ztext in zip(xs, ys, zs):
-                if not np.isnan(ztext) and xtext % 360 > bound_w % 360 and xtext % 360 < bound_e % 360 and\
-                        ytext > bound_s and ytext < bound_n:
-                    square_color = cmap(norm(ztext))
-                    square_brightness = np.mean(
-                        square_color[:3]) * square_color[3]
-                    text_color = 'k' if square_brightness > 0.5 else 'w'
-                    self.ax.text(xtext, ytext, ztext.astype(int), ha='center', va='center', fontsize=fs,
-                                 color=text_color, alpha=0.8, transform=ccrs.PlateCarree(), zorder=2)
-
-        # --------------------------------------------------------------------------------------
-
-        # Phantom legend
-        handles = []
-        for _ in range(10):
-            handles.append(mlines.Line2D(
-                [], [], linestyle='-', label='', lw=0))
-        l = self.ax.legend(handles=handles, loc='upper left',
-                           fancybox=True, framealpha=0, fontsize=11.5)
-        plt.draw()
-
-        # Get the bbox
-        try:
-            bb = l.legendPatch.get_bbox().inverse_transformed(self.fig.transFigure)
-        except:
-            bb = l.legendPatch.get_bbox().transformed(self.fig.transFigure.inverted())
-        bb_ax = self.ax.get_position()
-
-        # Define colorbar axis
-        cax = self.fig.add_axes(
-            [bb.x0 + 1.2 * bb.width, bb.y0 - .05 * bb.height, 0.015, bb.height])
-#        cbmap = mlib.cm.ScalarMappable(norm=norm, cmap=cmap)
-        cbar = self.fig.colorbar(cbmap, cax=cax, orientation='vertical',
-                                 ticks=clevs)
-
-        """
-        if len(prop['levels'])>2:
-            cax.yaxis.set_ticks(np.linspace(min(clevs),max(clevs),len(clevs)))
-            cax.yaxis.set_ticks(np.linspace(0,1,len(clevs)))
-            cax.yaxis.set_ticklabels(clevs)
-        else:
-            cax.yaxis.set_ticks(clevs)
-        """
-        cax.tick_params(labelsize=11.5)
-        cax.yaxis.set_ticks_position('left')
-
-        rect_offset = 0.0
-        if prop['cmap'] == 'category' and varname == 'vmax':
-            cax.yaxis.set_ticks(np.linspace(
-                min(clevs), max(clevs), len(clevs)))
-            cax.yaxis.set_ticklabels(clevs)
-            cax2 = cax.twinx()
-            cax2.yaxis.set_ticks_position('right')
-            cax2.yaxis.set_ticks((np.linspace(0, 1, len(clevs))[
-                                 :-1] + np.linspace(0, 1, len(clevs))[1:]) * .5)
-            cax2.set_yticklabels(
-                ['TD', 'TS', 'Cat-1', 'Cat-2', 'Cat-3', 'Cat-4', 'Cat-5'], fontsize=11.5)
-            cax2.tick_params('both', length=0, width=0, which='major')
-            cax.yaxis.set_ticks_position('left')
-
-            rect_offset = 0.7
-        if varname == 'date':
-            cax.set_yticklabels(
-                [f'{mdates.num2date(i):%b %-d}' for i in clevs], fontsize=11.5)
-
-        rectangle = mpatches.Rectangle((bb.x0, bb.y0 - 0.1 * bb.height), (2 + rect_offset) * bb.width, 1.1 * bb.height,
-                                       fc='w', edgecolor='0.8', alpha=0.8,
-                                       transform=self.fig.transFigure, zorder=3)
-        self.ax.add_patch(rectangle)
-
-        # --------------------------------------------------------------------------------------
-
-        # Add left title
-        try:
-            self.ax.set_title(prop['title_L'], loc='left',
-                              fontsize=17, fontweight='bold')
-        except:
-            pass
-
-        # Add right title
-        try:
-            self.ax.set_title(prop['title_R'], loc='right', fontsize=15)
-        except:
-            pass
-
-        # --------------------------------------------------------------------------------------
-
-        # Add plot credit
-        text = self.plot_credit()
-        self.add_credit(text)
-
-        # Return axis if specified, otherwise display figure
-        return self.ax
+    def plot_gridded_wrapper(self, *args, **kwargs):
+        return self.plot_gridded(*args, **kwargs)
 
     def plot_summary(self, storms, forecasts, shapefiles, valid_date, domain, ax=None, save_path=None, two_prop={}, invest_prop={}, storm_prop={}, cone_prop={}, map_prop={}):
         r"""
